+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/impl/module/FlattenedOptionalModule.java
@@ -16,44 +16,29 @@
 
 package com.github.victools.jsonschema.generator.impl.module;
 
-import com.fasterxml.classmate.ResolvedType;
-import com.github.victools.jsonschema.generator.MemberScope;
-import com.github.victools.jsonschema.generator.Module;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfigBuilder;
-import java.util.Collections;
-import java.util.List;
 import java.util.Optional;
 
 /**
  * Default module being included if {@code Option.FLATTENED_OPTIONALS} is enabled.
  */
-public class FlattenedOptionalModule implements Module {
-
-    private static boolean isOptional(ResolvedType type) {
-        return type.getErasedType() == Optional.class;
-    }
+public class FlattenedOptionalModule extends FlattenedWrapperModule<Optional> {
 
     /**
-     * Determine the type of the item/component wrapped in the given {@link Optional} type.
-     *
-     * @param fieldOrMethod reference to the method/field (which is being ignored here)
-     * @return the wrapped component type (or null if it is not an {@link Optional} (sub) type
+     * Constructor declaring {@link Optional} as the target type to unwrap.
      */
-    private List<ResolvedType> resolveOptionalComponentType(MemberScope<?, ?> fieldOrMethod) {
-        return FlattenedOptionalModule.isOptional(fieldOrMethod.getType())
-                ? Collections.singletonList(fieldOrMethod.getTypeParameterFor(Optional.class, 0))
-                : null;
+    public FlattenedOptionalModule() {
+        super(Optional.class);
     }
 
     @Override
     public void applyToConfigBuilder(SchemaGeneratorConfigBuilder builder) {
+        super.applyToConfigBuilder(builder);
         builder.forFields()
-                .withTargetTypeOverridesResolver(this::resolveOptionalComponentType)
-                // need to getDeclaredType() to avoid the above override
-                .withNullableCheck(field -> FlattenedOptionalModule.isOptional(field.getDeclaredType()) ? Boolean.TRUE : null);
+                // need to getDeclaredType() to avoid the target-type-override applied by generic wrapper module
+                .withNullableCheck(field -> this.hasMemberWrapperType(field) ? Boolean.TRUE : null);
         builder.forMethods()
-                .withTargetTypeOverridesResolver(this::resolveOptionalComponentType)
-                // need to getDeclaredType() to avoid the above override
-                .withNullableCheck(method -> FlattenedOptionalModule.isOptional(method.getDeclaredType()) ? Boolean.TRUE : null);
+                // need to getDeclaredType() to avoid the target-type-override applied by generic wrapper module
+                .withNullableCheck(method -> this.hasMemberWrapperType(method) ? Boolean.TRUE : null);
     }
 }